requirements.txt
pytorch
torchvision
tqdm
nltk
pillow
